Update to add source title to meta data
