+++ prisma/schema.prisma
@@ -1,4 +1,3 @@
-
 generator client {
   provider = "prisma-client-js"
 }
@@ -10,13 +9,14 @@ datasource db {
 }
 
 model Upload {
-  id         String   @id @default(cuid())
-  provider   String // can be 'mux' or 'cloudflare'
-  uploadLink String   @db.VarChar(2048)
-  assetId    String
-  uploadId   String
-  playbackId String?
-  status     String   @default("unknown")
-  createdAt  DateTime @default(now())
-  updatedAt  DateTime @updatedAt
+  id          String   @id @default(cuid())
+  provider    String // can be 'mux' or 'cloudflare'
+  uploadLink  String   @db.VarChar(2048)
+  assetId     String
+  uploadId    String
+  playbackId  String?
+  status      String   @default("unknown")
+  sourceTitle String?  @db.VarChar(2048) @default("ScreenLink Recording")
+  createdAt   DateTime @default(now())
+  updatedAt   DateTime @updatedAt
 }
